Agrego alertas

+++ app.js
@@ -28,16 +28,17 @@ localStorage.setItem("CatalogoCompleto", todosLosProductosStr);
 //Se declaran variables
 
 let usuario;
-let carrito = JSON.parse (localStorage.getItem('PeliculasEnCarrito')) || [];
+let carrito = JSON.parse(localStorage.getItem('ProductosEnCarrito')) || [];
 const body = document.querySelector("body");
 const botonCarrito = document.querySelector("#botonCarrito");
 botonCarrito.addEventListener("click", () => {
-    carrito.length != 0 &&  abrirCarrito();
-    })
+    carrito.length != 0 && abrirCarrito();
+})
 
 
 
 //Se declaran funciones
+
 
 function validarUsuario(user) {
     let saludo = document.querySelector("#userName");
@@ -52,6 +53,7 @@ function ingresarUsuario() {
     let nombre;
     let pantallaIngreso = document.querySelector("#bienvenida");
     pantallaIngreso.style.display = "block";
+    body.style.overflow = "hidden";
     let campo = document.querySelector("#nombreUsuario");
     campo.addEventListener("input", (e) => {
         nombre = e.target.value;
@@ -63,8 +65,8 @@ function ingresarUsuario() {
         nombre != undefined ? nombre : nombre = "usuario";
         validarUsuario(nombre);
     }
-    }
-    
+}
+
 
 function setButton(inBtn, addedClass, newId, container, action, reference) {
     let btn = document.createElement("button");
@@ -109,9 +111,17 @@ function mostrarProductos() {
     })
 }
 
+
 function mostrarInfo(array, indice) {
     const item = array[indice];
-    let {nombre, marca, modelo, descripcion, precio, foto} = item;
+    let {
+        nombre,
+        marca,
+        modelo,
+        descripcion,
+        precio,
+        foto
+    } = item;
     let modalInfo = document.querySelector("#modalInfoProductos");
     let itemInfo = document.createElement("div");
     itemInfo.classList.add("modalInfoContent");
@@ -130,34 +140,31 @@ function mostrarInfo(array, indice) {
     </div>`
     modalInfo.appendChild(itemInfo);
 
+
     let productoEncontradoCarrito = carrito.findIndex((elemento) => {
         return elemento.nombre === item.nombre
     });
+
     let btnCartText;
     productoEncontradoCarrito === -1 ? btnCartText = "Agregar al carrito" : btnCartText = "Quitar del carrito";
-
-    //let btnCartText;
-    //if (carrito.includes(item)) {
-    //    btnCartText = "Quitar del carrito"
-    //} else {
-    //    btnCartText = "Agregar al carrito"
-    //}
 
     setButton(btnCartText, "botonModal", "btnCart", itemInfo, agregarCarrito, item);
     setButton("Volver", "botonModal", "btnVolver", itemInfo, cerrarModal, modalInfo);
 
-    //setButton(btnCartText, "botonModal", itemInfo, agregarCarrito, item);
-    //setButton("Volver", "botonModal", itemInfo, cerrarModal, modalInfo);
-
     modalInfo.style.display = "block";
     body.style.overflow = "hidden";
-
 }
+
+mostrarProductos();
+
+mostrarNotificacion(textoNotificacion);
+
 
 function agregarCarrito(item) {
     let productoEncontrado = carrito.findIndex((elemento) => {
         return elemento.nombre === item.nombre
     });
+
     productoEncontrado === -1 ? carrito.push(item) : carrito.splice(productoEncontrado, 1);
 
     const carritoStr = JSON.stringify(carrito);
@@ -167,23 +174,22 @@ function agregarCarrito(item) {
 
     let textoBoton = document.querySelector("#btnCart");
     let btnCartText;
-    carrito.includes(item) ? btnCartText = "Quitar del carrito" : btnCartText = "Agregar al carrito";
+        if (carrito.includes(item)) {
+        btnCartText = "Agregar al carrito"
+        textoNotificacion = `${item.nombre} fue agregado al carrito`;
+        } else {
+        btnCartText = "Quitar del carrito"
+        textoNotificacion = `${item.nombre} fue quitado del carrito`;
+    }
+    
     textoBoton.innerHTML = btnCartText;
+
+    mostrarNotificacion(textoNotificacion);
 }
-    //if (productoEncontrado === -1) {
-       // carrito.push(item);
-       // const carritoStr = JSON.stringify(carrito);
-       // localStorage.setItem("ProductosEnCarrito", carritoStr);
-   // } else {
-      //  carrito.splice(productoEncontrado, 1);
-      //  const carritoStr = JSON.stringify(carrito);
-    //   localStorage.setItem("ProductosEnCarrito", carritoStr);
-    //}
-   // modificarContadorCarrito();
 
-//}
 
-function modificarContadorCarrito () {
+
+function modificarContadorCarrito() {
     let carritoContainer = document.querySelector("#carrito");
     let contadorCarrito = document.createElement("p");
     carritoContainer.innerHTML = ""
@@ -194,12 +200,13 @@ function modificarContadorCarrito () {
 }
 
 function abrirCarrito() {
+    let total = 0
     let modalCart = document.querySelector("#modalCart")
     let modalCarrito = document.querySelector("#modalCarrito");
     modalCarrito.innerHTML = ""
     if (carrito.length > 0) {
         carrito.forEach((producto) => {
-            total = producto.precio;
+            total = total + producto.precio;
             let modalContent = document.createElement("div");
             modalContent.classList.add("descripcionProducto");
             modalContent.innerHTML = `<img src="./img/${producto.miniatura}.webp" alt="">
@@ -209,7 +216,7 @@ function abrirCarrito() {
         })
         let montoTotal = document.createElement("div");
         montoTotal.classList.add("montoTotal");
-        montoTotal.innerHTML= "";
+        montoTotal.innerHTML = "";
         montoTotal.innerHTML = `<h4>Total de la compra: $${total}</h4>`
         modalCarrito.appendChild(montoTotal)
 
@@ -221,21 +228,52 @@ function abrirCarrito() {
         modalCarrito.appendChild(acciones)
     }
     modalCart.style.display = "block";
-    body.style.overflow = "hidden";  
+    body.style.overflow = "hidden";
 }
 
-function finalizarCompra(){
-    modalCarrito.innerHTML = "";
-    carrito = [];
-    modificarContadorCarrito();
-    modalCarrito.innerHTML = `<h3>¡Gracias por su compra!</h3> 
-    <button onClick="cerrarModal(modalCart)">Aceptar</button>`
-    const carritoStr = JSON.stringify(carrito);
-    localStorage.setItem("ProductosEnCarrito", carritoStr);
+function mostrarNotificacion(notificacion) {
+    Toastify({
+        text: notificacion,
+        duration: 2000,
+        gravity: "bottom",
+        className: "toastifyNotification",
+        style: {
+            background: "linear-gradient(to right, #4741A6, #9bbbfce5)",
+        }
+    }).showToast();
 }
+
+function finalizarCompra() {
+    modalCarrito.innerHTML = "";
+    cerrarModal(modalCart);
+
+    Swal.fire({
+        title: 'Estás a un paso de disfrutar del mejor producto',
+        text: "¿Confirmar compra?",
+        icon: 'info',
+        showCancelButton: true,
+        confirmButtonColor: '#3085d6',
+        cancelButtonColor: '#d33',
+        confirmButtonText: 'Comprar',
+        cancelButtonText: 'Volver',
+    }).then((result) => {
+        if (result.isConfirmed) {
+            Swal.fire(
+                'Compra realizada',
+                'Que disfrutes de tu producto',
+                'success'
+            )
+            carrito = [];
+            modificarContadorCarrito();
+            const carritoStr = JSON.stringify(carrito);
+            localStorage.setItem("ProductosEnCarrito", carritoStr);
+        }
+    })
+}
+
 
 // Fin de funciones
 
 ingresarUsuario();
 mostrarProductos();
-modificarContadorCarrito ();
+modificarContadorCarrito();
